requirements.txt
# Requirements for sprite sheet corpus fetcher
datasets>=2.14.0
huggingface-hub>=0.19.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
